Task/simplify build scripts

+++ scripts/lc-builds/toss3_hipcc.sh
@@ -1,13 +1,26 @@
 #!/usr/bin/env bash
 
 ###############################################################################
-# Copyright (c) 2016-20, Lawrence Livermore National Security, LLC
+# Copyright (c) 2016-21, Lawrence Livermore National Security, LLC
 # and RAJA project contributors. See the RAJA/COPYRIGHT file for details.
 #
 # SPDX-License-Identifier: (BSD-3-Clause)
 ###############################################################################
 
-BUILD_SUFFIX=lc_toss3-hipcc-3.10.0
+if [ "$1" == "" ]; then
+  echo
+  echo "You must pass a compiler version number to script. For example,"
+  echo "    toss3_hipcc.sh 4.0.1"
+  exit
+fi
+
+COMP_VER=$1
+
+BUILD_SUFFIX=lc_toss3-hipcc-${COMP_VER}
+
+echo
+echo "Creating build directory ${BUILD_SUFFIX} and generating configuration in it"
+echo
 
 rm -rf build_${BUILD_SUFFIX} >/dev/null
 mkdir build_${BUILD_SUFFIX} && cd build_${BUILD_SUFFIX}
@@ -17,10 +30,10 @@ mkdir build_${BUILD_SUFFIX} && cd build_${BUILD_SUFFIX}
 
 cmake \
   -DCMAKE_BUILD_TYPE=Release \
-  -DHIP_ROOT_DIR="/opt/rocm-3.10.0/hip" \
-  -DHIP_CLANG_PATH=/opt/rocm-3.10.0/llvm/bin \
-  -DCMAKE_C_COMPILER=/opt/rocm-3.10.0/llvm/bin/clang \
-  -DCMAKE_CXX_COMPILER=/opt/rocm-3.10.0/llvm/bin/clang++ \
+  -DHIP_ROOT_DIR="/opt/rocm-${COMP_VER}/hip" \
+  -DHIP_CLANG_PATH=/opt/rocm-${COMP_VER}/llvm/bin \
+  -DCMAKE_C_COMPILER=/opt/rocm-${COMP_VER}/llvm/bin/clang \
+  -DCMAKE_CXX_COMPILER=/opt/rocm-${COMP_VER}/llvm/bin/clang++ \
   -C ../host-configs/lc-builds/toss3/hip.cmake \
   -DENABLE_HIP=ON \
   -DENABLE_OPENMP=OFF \
